+++ dataworkspace/dataworkspace/tests/core/test_views.py
@@ -6,8 +6,6 @@ from django.contrib.auth.models import Permission
 from django.test import override_settings, Client
 
 from django.urls import reverse
-from django.shortcuts import render
-from waffle.models import Switch
 
 from dataworkspace.tests.common import (
     BaseTestCase,
@@ -158,25 +156,10 @@ def test_footer_links(request_client):
 
 
 @pytest.mark.parametrize(
-    "request_client, expected_template",
-    (("client", "tools-unauthorised.html"), ("staff_client", "tools.html")),
-    indirect=["request_client"],
-)
-def test_tools_only_shown_for_users_with_permissions(request_client, expected_template):
-    with mock.patch(
-        'dataworkspace.apps.applications.views.render', wraps=render
-    ) as renderer:
-        response = request_client.get(reverse("applications:tools"))
-
-    assert response.status_code == 200
-    assert renderer.call_args[0][1] == expected_template
-
-
-@pytest.mark.parametrize(
     "has_quicksight_access, expected_href, expected_text",
     (
-        (True, "https://quicksight", "Launch AWS QuickSight"),
-        (False, "/support-and-feedback/", "Request access"),
+        (True, "https://quicksight", "Open AWS QuickSight"),
+        (False, "/support-and-feedback/", "Request access to AWS QuickSight"),
     ),
 )
 @override_settings(QUICKSIGHT_SSO_URL='https://quicksight')
@@ -192,9 +175,6 @@ def test_quicksight_link_only_shown_to_user_with_permission(
         user.user_permissions.add(perm)
     user.save()
     client = Client(**get_http_sso_data(user))
-
-    switch = Switch(name='enable_quicksight_button', active=True)
-    switch.save()
 
     response = client.get(reverse("applications:tools"))